New UART Protocol + Message Queues Tested
- New UART protocol + message queues tested on Standard
- CAN protocol fix applied
- Current code idles until receives input from CV

Sid Prabhakaran
5/09/2021

+++ Core/Inc/information/uart_protocol.hpp
@@ -17,7 +17,7 @@ extern QueueHandle_t gimbalQueue;
 
 struct gimbMessage {
     uint8_t prefix;
-    int16_t disp[2];
+    float disp[2];
 };
 
 typedef enum {
